% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{get_edi_file}
\alias{get_edi_file}
\title{Download EDI Package Files
Download the latest revision of files in an EDI package}
\usage{
get_edi_file(pkg_id, fnames, verbose = TRUE)
}
\arguments{
\item{pkg_id}{The EDI Package ID}

\item{fnames}{A vector of file names in the package to download. Supports regex.}

\item{verbose}{If TRUE, display descriptive messages.}
}
\value{
a list of dataframes
}
\description{
Download EDI Package Files
Download the latest revision of files in an EDI package
}
