% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{get_edi_ids}
\alias{get_edi_ids}
\title{Get EMP EDI Package Names
Gives package ID for EMP data}
\usage{
get_edi_ids(program)
}
\arguments{
\item{program}{The program to retrieve the EDI ID for. Must be one of `r knitr::combine_words(programs, and = 'or ')`.}
}
\description{
Get EMP EDI Package Names
Gives package ID for EMP data
}
